Add functionality to remove a selected number of items from the cart

+++ backend/routes/products.py
@@ -6,7 +6,7 @@ from .auth import get_current_user
 
 router = APIRouter(prefix="/products", tags=["Products"])
 
-@router.get("/", response_model=list[ProductPublic], status_code=status.HTTP_200_OK)
+@router.get("/", response_model=list[Product], status_code=status.HTTP_200_OK)
 def get_products(session: Session = Depends(get_session)):
   products = session.exec(
       select(Product)
@@ -104,8 +104,10 @@ def remove_from_cart(id: str,
 
   if cart_item is None:
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Product not found")
-
-  session.delete(cart_item)
+  if cart_item.quantity > 0 and cart_item.quantity >= quantity:
+    cart_item.quantity -= quantity
+    session.add(cart_item)
+  if cart_item.quantity == 0: session.delete(cart_item)
   session.commit()
 
   return {"data": "Removed item from the cart"}
